+++ backend/canisters/group_index/impl/src/updates/c2c_update_community.rs
@@ -39,6 +39,31 @@ fn c2c_update_community_impl(args: Args, state: &mut RuntimeState) -> Response {
             args.gate,
         );
         Success
+    } else if let Some(community) = state.data.private_communities.get(&community_id) {
+        if state.data.public_group_and_community_names.is_name_taken(&args.name) {
+            return NameTaken;
+        }
+
+        let date_created = community.created();
+
+        state.data.private_communities.delete(&community_id);
+        state
+            .data
+            .public_group_and_community_names
+            .insert(&args.name, community_id.into());
+
+        state.data.public_communities.add(
+            community_id,
+            args.name,
+            args.description,
+            args.avatar_id,
+            args.banner_id,
+            args.gate,
+            args.primary_language,
+            args.channel_count,
+            date_created,
+        );
+        Success
     } else {
         CommunityNotFound
     }