+++ backend/canisters/group_index/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Improve upgrade version check to support multiple active versions ([#4215](https://github.com/open-chat-labs/open-chat/pull/4215))
+- Allow making private communities public ([#4217](https://github.com/open-chat-labs/open-chat/pull/4217))
 
 ## [[2.0.811](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.811-group_index)] - 2023-08-21